create a separate file for the LinkedList class
